requirements.txt
python-dotenv
pyautogui
crewai
portkey-ai
langchain_openai
langchain
crewai[tools]